requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.6
django-filter==25.1
django-jazzmin==3.0.1
django-widget-tweaks==1.5.0
idna==3.10
pillow==11.1.0
requests==2.32.3
slugify==0.0.1
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
